refactor: Update import paths for handCalculation and randomDice, removing the helpers directory
