+++ src/context/logic/handCalc/countDice.test.ts
