+++ src/context/logic/handCalc/countDice.ts
