+++ src/context/logic/handCalc/evaluateHandDices.ts
