+++ src/context/logic/handCalc/patterns.ts
