--- src/helpers/index.ts
@@ -1,2 +1,0 @@
-export { handCalculation } from './handCalc'
-export { randomDice } from './randomDice'
